backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow>=10.0.0
PyMySQL==1.1.2
cryptography>=42.0.0
pycryptodome>=3.20.0
python-dotenv==1.2.1
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
requests==2.32.3
django-cors-headers==4.9.0


